worker/requirements.txt
redis==5.2.1
boto3==1.36.26
opencv-python-headless==4.11.0.86
rembg==2.0.63
Pillow==11.1.0
onnxruntime==1.20.1
